move LLMConfig within greylitlm module
